Adjust terminal highlighting colors for code.
Error         This wasn't standing out enough, emboldened.

TODO          Was showing up as white on light grey, now on red.

ColorColumn   The `ctermbg` was simply set to the regular background and the
              foreground highlighting was a middle-grey.  This produced
              odd-looking off-color characters that seemed like some sort of
              error on the occasional ragged line.  Adjusted `ctermbg` to be
              similar to the guibg, and disabled the foreground highlighting.
              If you want to disable this, you can do so with `set colorcolumn=0`

CursorColumn  Was set to no highlighting.  This is a useful feature for
              whitespace-sensitive languages like Python or CoffeeScript, and if
              you want to disable it you can do so with `set nocursorcolumn`

CursorLine    Was set to no highlighting.  This is a useful feature for code in
              general.  If you want to disable it you can do so with
              `set nocursorline`